requirements.txt
blinker==1.4
selenium-wire==5.1.0
webdriver-manager==4.0.1